# Only clone the public properties for the exported component